3. MLE/1. Baseline model/requirements.txt
Babel==2.13.1
CacheControl==0.13.1
ConfigUpdater==3.2
Deprecated==1.2.14
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask==2.2.5
GitPython==3.1.42
Jinja2==3.1.2
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
PyJWT==2.8.0
PyYAML==6.0.1
Pygments==2.17.2
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
SQLAlchemy==1.4.52
SecretStorage==3.3.3
WTForms==3.0.1
Werkzeug==2.2.3
aiobotocore==2.12.0
aiohttp-retry==2.8.3
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.1.0
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-postgres==5.10.2
apache-airflow-providers-sqlite==3.5.0
apache-airflow-providers-telegram==4.6.0
apache-airflow==2.8.4
apispec==6.3.0
appdirs==1.4.4
argcomplete==3.1.6
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.1.0
backoff==2.2.1
billiard==4.2.0
blinker==1.7.0
boto3==1.34.51
botocore==1.34.51
build==1.0.3
cachelib==0.9.0
catboost==1.2.3
category-encoders==2.6.3
cattrs==23.2.3
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.1
configobj==5.0.8
connexion==2.14.2
contourpy==1.2.0
crashtest==0.4.1
cron_descriptor==1.4.0
croniter==2.0.2
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dictdiffer==0.9.0
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
distro==1.9.0
dnspython==2.4.2
docutils==0.20.1
dpath==2.1.6
dulwich==0.21.6
dvc-data==3.14.0
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.1
dvc-s3==3.0.1
dvc-studio-client==0.20.0
dvc-task==0.3.0
dvc==3.48.1
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.49.0
frozenlist==1.4.0
fsspec==2024.2.0
funcy==2.0
gitdb==4.0.11
google-re2==1.1
googleapis-common-protos==1.61.0
grandalf==0.8
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.3
gto==1.7.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
installer==0.7.0
ipykernel==6.29.3
ipython==8.22.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.3.2
jsonschema-specifications==2023.11.2
jsonschema==4.20.0
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.5
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
markdown-it-py==3.0.0
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
matplotlib==3.8.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
omegaconf==2.3.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.11.2
patsy==0.5.6
pendulum==2.1.2
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.19.0
pluggy==1.3.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
poetry==1.7.1
prison==0.2.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==2.0.0
pygit2==1.14.1
pygtrie==2.5.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
python-telegram-bot==21.5
pytz==2023.3.post1
pytzdata==2020.1
pyzmq==25.1.2
rapidfuzz==3.5.2
referencing==0.31.1
requests-toolbelt==1.0.0
requests==2.31.0
rfc3339-validator==0.1.4
rich-argparse==1.4.0
rich==13.7.0
rpds-py==0.13.2
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.6
s3fs==2024.2.0
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
scmrepo==3.2.0
semver==3.0.2
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.12
shtab==1.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
sqltrie==0.11.0
stack-data==0.6.3
statsmodels==0.14.1
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.3.0
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2023.11.29
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.25.0
voluptuous==0.14.2
wcwidth==0.2.13
wrapt==1.16.0
xattr==0.10.1
yarl==1.9.3
zc.lockfile==3.0.post1
zipp==3.17.0
